crypto/dimension-secp-256k/setup/pyproject.toml
[tool.poetry]
name = "ccsc-eth-nonce-reuse"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
web3 = "^5.29.0"
pycoin = "^0.92.20220213"
black = "^22.3.0"
pycryptodome = "^3.14.1"
ecdsa = "^0.17.0"
pysha3 = "^1.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
